+++ src/core/notifier/telegram/TelegramUserRepository.h
@@ -6,11 +6,11 @@
 
 class TelegramUserRepository {
 private:
-	std::map<std::string, int> db;
+	std::map<std::string, long> db;
 
 public:
-	std::map<std::string, int> getUsers();
-	void addUser(std::string username, int chat_id);
+	std::map<std::string, long> getUsers();
+	void addUser(std::string username, long chat_id);
 	int getUser(std::string username);
 	void removeUser(std::string username);
 };
